requirements.txt
# Core Python dependencies for this project.
# Install the matching PyTorch CUDA and DGL CUDA builds manually for GPU training.
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
tqdm==4.67.1
pypdf==6.9.1
pyyaml==6.0.3
pydantic==2.11.10
torchdata==0.7.1